+++ src/member/memberControl/FindPw.tsx
@@ -1,124 +1,195 @@
-import React from "react";
+import { useState, ChangeEvent, FormEvent } from "react";
 import styled from "styled-components";
+import { IoIdCardOutline, IoMailOutline } from "react-icons/io5";
+import bgImage from "../../img/g4.webp";
 
-const Body = styled.body`
-  margin: 0;
-  font-family: "Segoe UI", sans-serif;
-  background-color: #2c2f38;
-  color: #ffffff;
+const Section = styled.section`
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 100vh;
-`;
-
-const Container = styled.div`
-  background-color: #3a3c42;
-  padding: 40px;
-  border-radius: 8px;
+  min-height: 100vh;
   width: 100%;
-  max-width: 400px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+  position: relative;
+  &::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: url(${bgImage}) no-repeat center center / cover;
+    z-index: -2;
+  }
+  &::after {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.6);
+    z-index: -1;
+  }
 `;
 
-const H2 = styled.h2`
-  text-align: center;
-  margin-bottom: 24px;
+const FormBox = styled.div`
+  position: relative;
+  width: 400px;
+  height: 450px;
+  background-color: transparent;
+  border: 2px solid rgba(255, 255, 255, 0.5);
+  border-radius: 20px;
+  backdrop-filter: blur(15px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
 `;
 
-const FindPwForm = styled.form`
+const Form = styled.form`
   display: flex;
   flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  height: 100%;
+
+  h2 {
+    font-size: 2em;
+    color: #fff;
+    text-align: center;
+    margin-top: 20px;
+  }
+`;
+
+const InputBox = styled.div`
+  position: relative;
+  width: 310px;
+  border-bottom: 2px solid #fff;
 
   label {
-    margin-top: 12px;
-    margin-bottom: 4px;
-    font-weight: bold;
+    position: absolute;
+    top: 50%;
+    left: 5px;
+    transform: translateY(-50%);
+    color: #fff;
+    font-size: 1em;
+    pointer-events: none;
+    transition: 0.5s;
+  }
+
+  input:focus ~ label,
+  input:valid ~ label {
+    top: -5px;
   }
 
   input {
-    padding: 10px;
-    border-radius: 4px;
+    width: 100%;
+    height: 50px;
+    background: transparent;
     border: none;
-    background-color: #565962;
-    color: white;
+    outline: none;
+    font-size: 1em;
+    padding: 0 15px 0 5px;
+    color: #fff;
   }
 
-  input::placeholder {
-    color: #cccccc;
-  }
-
-  button {
-    margin-top: 24px;
-    padding: 12px;
-    background-color: #1ea7fd;
-    border: none;
-    border-radius: 4px;
-    color: white;
-    font-size: 16px;
-    cursor: pointer;
-  }
-
-  button:hover {
-    background-color: #0e82d6;
+  svg {
+    position: absolute;
+    right: 8px;
+    color: #fff;
+    font-size: 1.2em;
+    top: 20px;
   }
 `;
 
-const LinkWrap = styled.div`
-  text-align: center;
+const Button = styled.button`
+  margin-bottom: 30px;
+  width: 100%;
+  height: 40px;
+  border-radius: 40px;
+  background: transparent;
+  border: 2px solid #fff;
+  color: #fff;
+  outline: none;
+  cursor: pointer;
+  font-size: 1em;
+  font-weight: 600;
+
+  &:hover {
+    background: #fff;
+    color: black;
+  }
+`;
+
+const ResultBox = styled.div`
   margin-top: 20px;
-
-  a {
-    color: #1ea7fd;
-    text-decoration: none;
-    font-size: 0.9rem;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
+  color: #fff;
+  text-align: center;
+  font-size: 1.1em;
 `;
 
-const FindPw = () => {
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+export default function FindPw() {
+  const [form, setForm] = useState({
+    id: "",
+    email: "",
+  });
+
+  const [result, setResult] = useState<string | null>(null);
+
+  const onChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const { id, value } = e.target;
+    setForm({ ...form, [id]: value });
+  };
+
+  const onSubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    console.log("비밀번호 찾기 요청됨!");
-    // 여기서 실제 API 호출하기?(수일PT에게 물어보자)
+
+    // 👉 실제 API 호출로 변경 가능 (지금은 예시)
+    if (form.id === "testuser123" && form.email === "test@email.com") {
+      setResult("등록된 이메일로 임시 비밀번호를 발송했습니다.");
+    } else {
+      setResult("일치하는 회원 정보가 없습니다.");
+    }
   };
 
   return (
-    <Body>
-      <Container>
-        <H2>비밀번호 찾기</H2>
-        <FindPwForm onSubmit={handleSubmit}>
-          <label htmlFor="userid">아이디</label>
-          <input
-            type="text"
-            id="userid"
-            name="userid"
-            placeholder="아이디를 입력하세요"
-            required
-          />
+    <Section>
+      <FormBox>
+        <Form onSubmit={onSubmit}>
+          <h2>비밀번호 찾기</h2>
 
-          <label htmlFor="email">이메일</label>
-          <input
-            type="email"
-            id="email"
-            name="email"
-            placeholder="이메일을 입력하세요"
-            required
-          />
+          <InputBox>
+            <input
+              type="text"
+              id="id"
+              required
+              value={form.id}
+              onChange={onChange}
+            />
+            <label htmlFor="id">아이디</label>
+            <IoIdCardOutline />
+          </InputBox>
 
-          <button type="submit">인증 메일 보내기</button>
-        </FindPwForm>
+          <InputBox>
+            <input
+              type="email"
+              id="email"
+              required
+              value={form.email}
+              onChange={onChange}
+            />
+            <label htmlFor="email">이메일</label>
+            <IoMailOutline />
+          </InputBox>
 
-        {/* {여기엔 로그인, 회원가입 링크 추가해도 좋다는디요?} */}
-        <LinkWrap>
-          <a href="/login">로그인으로 돌아가기</a>
-        </LinkWrap>
-      </Container>
-    </Body>
+          <Button type="submit">비밀번호 찾기</Button>
+
+          {result && (
+            <ResultBox>
+              <p>{result}</p>
+            </ResultBox>
+          )}
+        </Form>
+      </FormBox>
+    </Section>
   );
-};
-
-export default FindPw;
+}
